+++ mxslc/mxslc/Expressions/__init__.py
@@ -12,3 +12,4 @@ from .UnaryExpression import UnaryExpression
 from .NodeConstructor import NodeConstructor
 from .IndexingExpression import IndexingExpression
 from .TernaryRelationalExpression import TernaryRelationalExpression
+from .VariableDeclarationExpression import VariableDeclarationExpression