+++ mxslc/mxslc/Parameter.py
@@ -11,10 +11,11 @@ class Parameter:
     """
     Represents a parameter to a function or constructor call.
     """
-    def __init__(self, identifier: Token, data_type: Token | DataType, default_value: Expression = None):
+    def __init__(self, identifier: Token, data_type: Token | DataType, default_value: Expression = None, is_out=False):
         self.__identifier = identifier
         self.__data_type = DataType(data_type)
         self.__default_value = default_value
+        self.__is_out = is_out
 
     @property
     def name(self) -> str:
@@ -32,12 +33,23 @@ class Parameter:
     def default_value(self) -> Expression | None:
         return self.__default_value
 
+    @property
+    def is_in(self) -> bool:
+        return not self.__is_out
+
+    @property
+    def is_out(self) -> bool:
+        return self.__is_out
+
     def init_default_value(self) -> None:
         if self.default_value:
             self.default_value.init(self.data_type)
 
     def __str__(self) -> str:
-        output = f"{self.data_type} {self.name}"
+        output = ""
+        if self.is_out:
+            output += "out "
+        output += f"{self.data_type} {self.name}"
         if self.default_value:
             output += f" = {self.default_value}"
         return output
@@ -54,6 +66,14 @@ class ParameterList:
             self.__params: list[Parameter] = params
         else:
             self.__params: list[Parameter] = []
+
+    @property
+    def ins(self) -> ParameterList:
+        return ParameterList([p for p in self if p.is_in])
+
+    @property
+    def outs(self) -> ParameterList:
+        return ParameterList([p for p in self if p.is_out])
 
     def instantiate_templated_parameters(self, template_type: DataType) -> ParameterList:
         return ParameterList([