+++ mxslc/mxslc/Statements/ExpressionStatement.py
@@ -15,3 +15,8 @@ class ExpressionStatement(Statement):
     def execute(self) -> None:
         node = self.__expr.init_evaluate()
         self._add_attributes_to_node(node)
+        if node.is_null_node:
+            node.remove()
+
+    def __str__(self) -> str:
+        return f"{self.__expr};"