+++ mxslc/mxslc/node_utils.py
@@ -11,12 +11,15 @@ def create(category: str, data_type: DataType | str) -> Node:
     return state.get_graph().add_node(category, data_type)
 
 
-def constant(value: Uniform) -> Node:
+def constant(value: Uniform = None, data_type: DataType = None) -> Node:
     """
     Add constant node to the current states graph element.
     """
-    node = create("constant", type_of(value))
-    node.set_input("value", value)
+    node = create("constant", data_type or type_of(value))
+    if value is not None:
+        node.add_input("value", value)
+    else:
+        node.add_input("value", data_type=data_type)
     return node
 
 
@@ -65,10 +68,6 @@ def convert(in_: Node, output_type: DataType) -> Node:
     """
     Add convert node to the current states graph element.
     """
-    unconvertable_types = [STRING, FILENAME, *SHADER_TYPES]
-    assert in_.data_type not in unconvertable_types
-    assert output_type not in unconvertable_types
-
     node = create("convert", output_type)
     node.set_input("in", in_)
     return node